Got the settings window to work
- Moved the settings window from "Windows" to "Tools" in the menu
- Hooked up the values in the settings to the actual in-game stats

+++ Assets/_Game/Scripts/SettingsManager.cs
@@ -9,6 +9,6 @@ public class SettingsManager : ScriptableObject
     [Header("Ship")]
     public float _throttlePower;
     public float _rotationPower;
-    public float _maxHealth;
+    public int _maxHealth;
 
 }
